skillflow-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
websockets>=13.0
opencv-python>=4.10.0.84
ffmpeg-python>=0.2.0
openai>=1.50.0
requests>=2.32.0
pydantic>=2.9.0
python-dotenv>=1.0.1
yt-dlp>=2024.10.22
pytest>=8.3.0
pytest-asyncio>=0.24.0
